re-init: bring files from archived repo's develop branch
Our long-standing work repo for this project is located at:
github.com/NASA-AMMOS/quickstart-ammos-smallsat-toolkit
which is a fork of:
https://github.com/aws-quickstart/quickstart-ammos-smallsat-toolkit

The above repo has now been archived in favor of:
https://github.com/aws-ia/cfn-ps-ammos-smallsat-toolkit

The two upstream repos share no commit history so we must manually copy
over all of the files from the now orphaned fork
(NASA-AMMOS-quickstart). This new fork should now be up-to-date although
most of this content was copied over blindly and may not perfectly fit
the new intended layout of this repo.